Make changes from feedback and update a few types
- The 2 entry points are now graphql-core-server and graphql-admin-server
- The user methods are now under src/db/user.ts
