+++ src/graphql/types/mutations/merchant-update-map-info.ts
@@ -1,6 +1,6 @@
 import { GT } from "../../index"
 
-import { updateMerchantMapInfo } from "../../db"
+import { updateMerchantMapInfo } from "../../../db/user"
 import UserDetailPayload from "../payloads/user-detail"
 import Username from "../scalars/username"
