+++ src/graphql/types/mutations/user-update-status.ts
@@ -1,5 +1,5 @@
 import { GT } from "../../index"
-import { updateUserAccountStatus } from "../../db"
+import { updateUserAccountStatus } from "../../../db/user"
 import AccountStatus from "../account-status"
 import UserDetailPayload from "../payloads/user-detail"
