+++ src/graphql/types/user-details.ts
@@ -4,6 +4,8 @@ import Date from "./scalars/date"
 import Phone from "./scalars/phone"
 import AccountLevel from "./account-level"
 import Username from "./scalars/username"
+import Coordinates from "./coordinates"
+import AccountStatus from "./account-status"
 
 const UserDetails = new GT.Object({
   name: "UserDetails",
@@ -19,6 +21,16 @@ const UserDetails = new GT.Object({
     },
     level: {
       type: AccountLevel,
+    },
+    status: {
+      type: AccountStatus,
+    },
+    title: {
+      type: GT.String,
+    },
+    coordinates: {
+      type: Coordinates,
+      resolve: (source) => source.coordinate,
     },
     createdAt: {
       type: GT.NonNull(Date),
