+++ src/schema.ts
@@ -367,15 +367,6 @@ UserSchema.statics.getUserByUsername = async function (username: string) {
   return user
 }
 
-UserSchema.statics.findOrCreate = async function (userInfo) {
-  let user
-  user = await this.findOne(userInfo)
-  if (!user) {
-    user = await this.create(userInfo)
-  }
-  return user
-}
-
 UserSchema.statics.getUserByAddress = async function ({ address }) {
   return await this.findOne({ "onchain.address": address })
 }
