+++ test/integration/02-user-wallet/02-index.spec.ts
@@ -1,7 +1,7 @@
 import { yamlConfig } from "src/config"
 import { getUserWallet } from "test/helpers"
 import { setAccountStatus } from "src/AdminOps"
-import { usernameExists } from "src/graphql/db"
+import { usernameExists } from "src/db/user"
 
 jest.mock("src/realtimePrice", () => require("test/mocks/realtimePrice"))
