+++ test/integration/entrypoint/graphql/admin/queries.spec.ts
@@ -4,7 +4,10 @@ import { User } from "src/schema"
 import { gqlAdminSchema } from "src/graphql/admin"
 
 beforeAll(async () => {
-  await User.findOrCreate({ username: "tester", phone: "+19876543210" })
+  let user = await User.findOne({ username: "tester", phone: "+19876543210" })
+  if (!user) {
+    user = await User.create({ username: "tester", phone: "+19876543210" })
+  }
 })
 
 describe("GraphQLQueryRoot", () => {
